deploy/requirement.txt
streamlit==1.26.0
pillow==9.4.0
pandas==2.1.1
opencv-python==4.7.0.72
fingerprint-enhancer
openpyxl
